refactor: standardize guess text in table
Current guess text is a one-to-one of whats in the spreadsheet. If the
initial parsing was without error, then that means the format was
concluded as fine, and we can just show a prettier version to look at.

This also sets a max possible width the guess can be, which can be hard
coded to futher standardize the look and feel of the app if so chosen.